input: gsl_ts: disable interrupts while i2c is in suspend
When the touchscreen wakes the system up, I2C ops fail sometimes if
the ISR runs before the device has properly resume.  The MSM I2C
adapter attempts to handle this case, but it's not reliable.

Add PM ops to enable/disable interrupts during suspend and resume.
This change ensures that the ISR never runs while the I2C adapter is
not ready.

Change-Id: I76986b333ca889c8ae7d0f742d5b77ffc168546e
Signed-off-by: Khalid Zubair <kzubair@cyngn.com>
Signed-off-by: Kiran Anto <kirananto@gmail.com>