input: misc: yas53x: Add sensors max delay time
Max delay is the delay between two sensor events
corresponding to the lowest frequency that this
sensor supports. When lower frequencies are requested
through batch()/setDelay() the events will be generated
at this frequency instead. It can be used by the
framework or applications to estimate when the batch
FIFO may be full.

Adapted from Daqing Chen <chendaqing@codeaurora.org> patch

Change-Id: I1a5ba7211aaa8441f49e7f6d2ab4e56c7ff83e35

Signed-off-by: Kiran Anto <kirananto@gmail.com>